+++ src/kimi_cli/tools/file/read.py
@@ -5,7 +5,7 @@ from kosong.tooling import CallableTool2, ToolError, ToolOk, ToolReturnType
 from pydantic import BaseModel, Field
 
 from kaos.path import KaosPath
-from kimi_cli.soul.runtime import BuiltinSystemPromptArgs
+from kimi_cli.soul.agent import BuiltinSystemPromptArgs
 from kimi_cli.tools.utils import load_desc, truncate_line
 
 MAX_LINES = 1000