+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Only write entries that are worth mentioning to users.
 ## [Unreleased]
 
 - Core: Fix field inheritance of agent spec files when using `extend`
+- Core: Support using MCP tools in subagents
+- Tool: Add `CreateSubagent` tool to create subagents dynamically (not enabled in default agent)
 
 ## [0.57] - 2025-11-20